[deprecation] Remove Cilium BGP load balancer support
