+++ deploy/chart/templates/clusterrole-rbac.yaml
@@ -31,12 +31,4 @@ rules:
   - apiGroups: [""]
     resources: ["services/status"]
     verbs: ["get", "watch", "list", "update", "patch"]
-{{- if .Values.sharedIPLoadBalancing }}
-  - apiGroups: ["cilium.io"]
-    resources: ["ciliumloadbalancerippools"]
-    verbs: ["get", "watch", "list", "update", "create", "delete"]
-  - apiGroups: ["cilium.io"]
-    resources: ["ciliumbgppeeringpolicies"]
-    verbs: ["get", "watch", "list", "create"]
-{{- end }}
 {{- end }}
