+++ docs/examples/README.md
@@ -13,7 +13,6 @@ This section provides working examples of common CCM configurations. Each exampl
    - Custom Health Checks
    - Firewalled Services
    - Session Affinity
-   - Shared IP Load-Balancing
    - Custom Node Selection
 
 3. **Frontend VPC NodeBalancer**
@@ -24,6 +23,7 @@ Note: To test UDP based NBs, one can use [test-server](https://github.com/rahula
 For testing these examples, see the [test script](https://github.com/linode/linode-cloud-controller-manager/blob/master/examples/test.sh).
 
 For more configuration options, see:
+
 - [Service Annotations](../configuration/annotations.md)
 - [LoadBalancer Configuration](../configuration/loadbalancer.md)
 - [Firewall Configuration](../configuration/firewall.md)
