--- e2e/bgp-test/lb-cilium-bgp/create-pod-service.yaml
@@ -1,24 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: test-pod-1
-  labels:
-    app: test-bgp
-spec:
-  containers:
-    - name: nginx
-      image: nginx:latest
-      ports:
-        - containerPort: 80
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: test-bgp-svc
-spec:
-  type: LoadBalancer
-  ports:
-    - port: 80
-      targetPort: 80
-  selector:
-    app: test-bgp
